+++ release/index.es6.js
@@ -33,7 +33,7 @@ export { SpinBox } from './ui/layout/SpinBox/SpinBox.es6.js';
 export { Navigation } from './ui/layout/Navigation/Navigation.es6.js';
 export { Details } from './ui/layout/Details/Details.es6.js';
 export { Image } from './ui/layout/Image/Image.es6.js';
-export { Pages, splitPagesItem } from './ui/layout/Pages/Pages.es6.js';
+export { Pages, splitPagesItem, titleContext, titleSeparatorContext } from './ui/layout/Pages/Pages.es6.js';
 export { AsyncSpin } from './ui/layout/AsyncSpin/AsyncSpin.es6.js';
 export { BadgeFooter } from './ui/layout/BadgeFooter/BadgeFooter.es6.js';
 export { Notification } from './ui/popups/Notification/Notification.es6.js';