+++ release/ui/layout/Pages/Pages.d.ts
@@ -1,3 +1,4 @@
+import { Context } from '@innet/jsx';
 import { type NavigationItemProps } from '../Navigation';
 export type PagesMenu = PagesItemProps[];
 export type PageAccessHandler = (nav: PagesItemProps) => any;
@@ -13,6 +14,10 @@ export interface PagesProps {
     navigation: PagesMenu;
     prefix?: string;
     handleAccess?: PageAccessHandler;
+    title?: string;
+    titleSeparator?: string;
 }
+export declare const titleContext: Context<string, string>;
+export declare const titleSeparatorContext: Context<string, string>;
 export declare function splitPagesItem(navigation: PagesMenu, prefix: string, handleAccess?: PageAccessHandler, parent?: any): [NavigationItemProps[], any];
-export declare function Pages({ navigation, prefix, handleAccess, }: PagesProps): any;
+export declare function Pages({ navigation, prefix, title, titleSeparator, handleAccess, }: PagesProps): any;