+++ release/ui/layout/Pages/Pages.es6.js
@@ -1,6 +1,6 @@
 import { __rest, __asyncGenerator, __await } from 'tslib';
 import { inject, routerContext } from '@innet/dom';
-import { useChildren } from '@innet/jsx';
+import { Context, useChildren } from '@innet/jsx';
 import '../AsyncSpin/index.es6.js';
 import '../Navigation/index.es6.js';
 import '../Page/index.es6.js';
@@ -8,6 +8,8 @@ import { AsyncSpin } from '../AsyncSpin/AsyncSpin.es6.js';
 import { DelayPage } from '../Page/Page.es6.js';
 import { Navigation } from '../Navigation/Navigation.es6.js';
 
+const titleContext = new Context();
+const titleSeparatorContext = new Context(' | ');
 function splitPagesItem(navigation, prefix, handleAccess, parent) {
     const menu = [];
     const pages = {
@@ -58,14 +60,17 @@ function splitPagesItem(navigation, prefix, handleAccess, parent) {
     }
     return [menu, pages];
 }
-function Pages({ navigation, prefix = '', handleAccess, }) {
+function Pages({ navigation, prefix = '', title, titleSeparator = ' | ', handleAccess, }) {
     const children = useChildren();
+    if (title) {
+        document.title = title;
+    }
     const [menu, pages] = splitPagesItem(navigation, prefix, handleAccess);
     const slots = ([{type:'slot',props:{name:'menu'},children:[{type:Navigation,props:{menu:menu}}]},{type:'slot',props:{name:'pages'},children:[pages]}]);
-    if (!prefix) {
-        return ({type:'slots',props:{from:slots},children:[children]});
-    }
-    return ({type:'slots',props:{from:slots},children:[{type:'context',props:{for:routerContext,set:(prefix.match(/\//g) || []).length + 1},children:[children]}]});
+    const result = prefix
+        ? ({type:'slots',props:{from:slots},children:[{type:'context',props:{for:routerContext,set:(prefix.match(/\//g) || []).length + 1},children:[children]}]})
+        : ({type:'slots',props:{from:slots},children:[children]});
+    return ({type:'context',props:{for:titleContext,set:title},children:[{type:'context',props:{for:titleSeparatorContext,set:titleSeparator},children:[result]}]});
 }
 
-export { Pages, splitPagesItem };
+export { Pages, splitPagesItem, titleContext, titleSeparatorContext };