+++ release/ui/layout/Pages/index.js
@@ -8,3 +8,5 @@ var Pages = require('./Pages.js');
 
 exports.Pages = Pages.Pages;
 exports.splitPagesItem = Pages.splitPagesItem;
+exports.titleContext = Pages.titleContext;
+exports.titleSeparatorContext = Pages.titleSeparatorContext;