+++ release/ui/typography/Title/Title.es6.js
@@ -1,8 +1,9 @@
 import { __rest } from 'tslib';
 import { style, useShow, useHidden } from '@innet/dom';
-import { useChildren } from '@innet/jsx';
+import { useChildren, useContext } from '@innet/jsx';
 import '../../layout/index.es6.js';
 import modules_f1b30145 from './Title.scss.es6.js';
+import { titleContext, titleSeparatorContext } from '../../layout/Pages/Pages.es6.js';
 import { Flex } from '../../layout/Flex/Flex.es6.js';
 
 const useStyle = style(modules_f1b30145);
@@ -11,9 +12,16 @@ function Title(_a = {}) {
     const children = useChildren();
     const show = useShow();
     const hide = useHidden();
+    const contextTitle = useContext(titleContext);
+    const titleSeparator = useContext(titleSeparatorContext);
     const styles = useStyle();
-    if (text !== undefined && h === 1) {
-        document.title = text;
+    if (h === 1 && (text !== undefined || contextTitle !== undefined)) {
+        if (contextTitle) {
+            document.title = !text || contextTitle === text ? contextTitle : `${contextTitle}${titleSeparator}${text}`;
+        }
+        else if (text) {
+            document.title = text;
+        }
     }
     return ({type:Flex,props:{element:`h${h}`,...props,class:() => [
             styles.root,